requirements.txt
dash==2.7.1
dash_bootstrap_components==1.3.0
pandas==1.5.0
Pillow==9.4.0
gunicorn
dash-tools
